Moving SlidesViewProps types to view file
