+++ packages/@coorpacademy-app-review/src/reducers/ui/index.ts
@@ -2,16 +2,13 @@ import {combineReducers} from 'redux';
 
 import currentSlideRef, {CurrentSlideRefState} from './current-slide-ref';
 import navigation, {NavigationState} from './navigation';
-import slides, {UISlidesState} from './slides';
 
 export type UIState = {
   currentSlideRef: CurrentSlideRefState;
   navigation: NavigationState;
-  slides: UISlidesState;
 };
 
 export default combineReducers({
   currentSlideRef,
-  navigation,
-  slides
+  navigation
 });
