--- packages/@coorpacademy-app-review/src/types/views/slides.ts
@@ -1,90 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import type {StoreState} from '../../reducers';
-import {Dispatchers} from '../../actions';
-
-const ICON_VALUES = {
-  right: 'right',
-  wrong: 'wrong',
-  'no-answer': 'no-answer'
-} as const;
-
-type IconValue = keyof typeof ICON_VALUES;
-
-type StepItem = {
-  current: boolean;
-  icon: IconValue;
-  value: string;
-};
-
-type CorrectionPopinProps = {
-  klf: {
-    label: string;
-    onClick: Function;
-    tooltip: string;
-  };
-  information: {
-    label: string;
-    message: string;
-  };
-  next: {
-    label: string;
-    ariaLabel: string;
-  };
-  successLabel: string;
-  failureLabel: string;
-};
-
-type SlidesViewStaticProps = {
-  header: {
-    mode: string;
-    skillName: string;
-    onQuitClick: Function;
-    'aria-label'?: string;
-    closeButtonAriaLabel: string;
-    steps: StepItem[];
-  };
-  stack: {
-    slides: StoreState['ui']['slides'];
-    validateButton: {
-      label: string;
-      disabled: boolean;
-      onClick: Function;
-    };
-    correctionPopinProps?: CorrectionPopinProps;
-    endReview: boolean;
-  };
-  reviewBackgroundAriaLabel?: string;
-  congratsProps?: {
-    'aria-label'?: string;
-    'data-name'?: string;
-    animationLottie: any;
-    title: string;
-    cardCongratsStar: any;
-    cardCongratsRank: any;
-    buttonRevising: {
-      'aria-label'?: string;
-      label: string;
-      onClick: Function;
-      type: string;
-    };
-    buttonRevisingSkill: {
-      'aria-label'?: string;
-      label: string;
-      onClick: Function;
-      type: string;
-    };
-  };
-};
-
-type SlidesViewDispatcherProps = Omit<Dispatchers, 'navigateTo' | 'navigateBack'>;
-
-type SlidesViewProps = SlidesViewStaticProps & SlidesViewDispatcherProps;
-
-export {
-  CorrectionPopinProps,
-  IconValue,
-  SlidesViewStaticProps,
-  SlidesViewDispatcherProps,
-  SlidesViewProps,
-  StepItem
-};
